+++ CHANGELOG.md
@@ -8,7 +8,6 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 
 - `value` method to replace the removed `getContent` method.
 - requires `League\Uri` package.
-- `UserInfo::new` named constructor
 - `Scheme::new` named constructor
 - `Host::new` named constructor
 - `Domain::new` named constructor
@@ -16,7 +15,10 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - `Query::new` named constructor
 - `Path::new` named constructor
 - `DataPath::new` named constructor
-- `UserInfo::withPass` modifier method
+- `UserInfo::withUsername` modifier method
+- `UserInfo::withPassword` modifier method
+- `UserInfo::getUsername`
+- `UserInfo::getPassword`
 
 ### Fixed
 
@@ -36,6 +38,8 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - Support for `__set_state`
 - Support for `float` type as possible argument for components
 - `UserInfo::withUserInfo` modifier method
+- `UserInfo::getUser` use instead `UserInfo::getUsername`
+- `UserInfo::getPass` use instead `UserInfo::getPassword`
 
 ## [2.4.1](https://github.com/thephpleague/uri-components/compare/2.4.0...2.4.1) - 2022-05-26